Add styling for each book list/info
